LayoutTests/imported/w3c/web-platform-tests/html/user-activation/activation-trigger-keyboard-escape.html

<!DOCTYPE html>
<html>
  <meta charset="utf-8" />
  <title>Tests missing user activation from a ESCAPE keyboard event.</title>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/interaction.html#activation-triggering-input-event"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body>
    <input type="text" autofocus />
  </body>
  <script>
    promise_test(async () => {
      const ESCAPE_KEY = "\uE00C";

      const keydown_event = getEvent("keydown");
      const keyup_event = getEvent("keyup");

      await test_driver.send_keys(document.body, ESCAPE_KEY);

      await keydown_event;
      assert_false(
        navigator.userActivation.isActive,
        "ESCAPE keydown event should not result in activation"
      );
      await consumeTransientActivation();

      await keyup_event;
      assert_false(
        navigator.userActivation.isActive,
        "ESCAPE keyup should have no activation after keydown consumption"
      );
    }, "Activation through ESCAPE keyboard event");
  </script>
</html>
